+++ internal/provider/provider.go
@@ -351,5 +351,6 @@ func (p *proxmoxProvider) DataSources(_ context.Context) []func() datasource.Dat
 func (p *proxmoxProvider) Resources(_ context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		nodefirewall.NewRulesResource,
+		nodefirewall.NewRuleResource,
 	}
 }